updated post test and renamed the package to restful.ecommerce
