+++ src/test/java/restful/ecommerce/TestDelayedAPIWithAwaitility.java
@@ -1,10 +1,7 @@
-package in.reqres;
+package restful.ecommerce;
 
 import static io.restassured.RestAssured.post;
 import static org.awaitility.Awaitility.await;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-import static io.restassured.RestAssured.put;
 import static org.hamcrest.Matchers.equalTo;
 
 import io.qameta.allure.Description;
@@ -21,7 +18,7 @@ import org.testng.annotations.Test;
  **/
 @Epic ("Rest Assured POC - Example Tests")
 @Feature ("Performing different API Tests using Rest-Assured")
-public class TestDelayedAPIWithAwaitility extends  AwaitilityBase {
+public class TestDelayedAPIWithAwaitility extends AwaitilityBase {
 
     @Test
     @Description ("Example Test for executing GET request using awaitility")
