+++ src/test/java/restful/ecommerce/TestDeleteRequests.java
@@ -13,7 +13,7 @@
         limitations under the License.
 */
 
-package in.reqres;
+package restful.ecommerce;
 
 import static io.restassured.RestAssured.given;
 
@@ -61,7 +61,8 @@ public class TestDeleteRequests {
     @Severity (SeverityLevel.NORMAL)
     @Story ("Execute Delete requests using rest-assured")
     public void deleteRequestTests (final int userId) {
-        given ().header ("x-api-key","reqres-free-v1").when ()
+        given ().header ("x-api-key", "reqres-free-v1")
+            .when ()
             .delete (URL + userId)
             .then ()
             .assertThat ()
