+++ src/test/java/restful/ecommerce/TestExpectations.java
@@ -1,4 +1,4 @@
-package in.reqres;
+package restful.ecommerce;
 
 import static io.restassured.RestAssured.expect;
 import static org.hamcrest.Matchers.equalTo;
@@ -23,7 +23,7 @@ public class TestExpectations {
     public void testBodyUsingExpect () {
         expect ().header ("x-api-key", "reqres-free-v1")
             .body ("data.email", equalTo ("janet.weaver@reqres.in"))
-            .body ("data.id", equalTo(2))
+            .body ("data.id", equalTo (2))
             .when ()
             .get ("https://reqres.in/api/users/2")
             .then ()
