+++ src/test/java/restful/ecommerce/TestPostWithRecords.java
@@ -1,0 +1,34 @@
+package restful.ecommerce;
+
+import static io.restassured.RestAssured.given;
+
+import io.restassured.http.ContentType;
+import org.testng.annotations.Test;
+
+public class TestPostWithRecords {
+    public record UserData(String name, String job) {
+    }
+
+    private static final String URL = "https://reqres.in";
+
+    @Test
+    public void testCreateUser () {
+
+        final UserData userData = new UserData ("Faisal", "QA");
+
+        final String response = given ().contentType (ContentType.JSON)
+            .header ("x-api-key", "reqres-free-v1")
+            .body (userData)
+            .when ()
+            .post (URL + "/api/users")
+            .then ()
+            .assertThat ()
+            .statusCode (201)
+            .extract ()
+            .response ()
+            .asString ();
+
+        System.out.println (response);
+    }
+
+}
