+++ test-suite/testng.xml
@@ -3,7 +3,7 @@
 <suite name="Rest-Assured API Test Suite">
     <listeners>
         <listener
-                class-name="in.reqres.utility.TestListener"/>
+                class-name="restful.ecommerce.utility.TestListener"/>
     </listeners>
     <suite-files>
         <!--        <suite-file path="reqrestestsuite.xml"/>-->
